pyproject.toml
[project]
name = "dbos"
dynamic = ["version"]
description = "Ultra-lightweight durable execution in Python"
authors = [
    {name = "DBOS, Inc.", email = "contact@dbos.dev"},
]
dependencies = [
    "pyyaml>=6.0.2",
    "jsonschema>=4.23.0",
    "alembic>=1.13.3",
    "typing-extensions>=4.12.2; python_version < \"3.10\"",
    "typer>=0.12.5",
    "jsonpickle>=3.3.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "python-dateutil>=2.9.0.post0",
    "fastapi[standard]>=0.115.2",
    "tomlkit>=0.13.2",
    "psycopg[binary]>=3.1", # Keep compatibility with 3.1--older Python installations/machines can't always install 3.2
    "docker>=7.1.0",
    "cryptography>=43.0.3",
    "rich>=13.9.4",
    "pyjwt>=2.10.1",
    "websockets>=15.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
dbos = "dbos.cli.cli:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"
version_format = "version:format_version"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
filter_files = true
atomic = true

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = "-s"
log_cli_format = "%(asctime)s [%(levelname)8s] (%(name)s:%(filename)s:%(lineno)s) %(message)s"
log_cli_level = "INFO"
log_cli = true

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "mypy>=1.15.0",
    "pytest-mock>=3.14.0",
    "types-PyYAML>=6.0.12.20240808",
    "types-jsonschema>=4.23.0.20240813",
    "black>=24.10.0",
    "pre-commit>=4.0.1",
    "isort>=5.13.2",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20240914",
    "httpx>=0.27.2",
    "pytz>=2024.2",
    "GitPython>=3.1.43",
    "confluent-kafka>=2.6.0",
    "types-confluent-kafka>=1.2.2",
    "flask>=3.0.3",
    "pytest-order>=1.3.0",
    "pdm-backend>=2.4.2",
    "pytest-asyncio>=0.25.0",
    "pyright>=1.1.398",
    "types-docker>=7.1.0.20241229",
]
